Corrige eInvalidInput al forzar ATTDISP (tipo int en vez de short)
El fix anterior (forzar ATTDISP=1 al arrancar CVENT/CVENTT) rompia el
comando entero: Application.SetSystemVariable("ATTDISP", 1) pasaba el valor
como int (32 bits), pero ATTDISP es un entero de 16 bits (short) a nivel
interno -el tipo tiene que coincidir exactamente, aunque el valor en si sea
valido-, asi que la llamada disparaba eInvalidInput sin capturar, tirando
abajo el comando desde la primera linea. Corregido forzando (short)1.

+++ ConductosPlugin/ConductosPlugin.cs
@@ -162,7 +162,10 @@ namespace ConductosPlugin
         /// dependa solo de su propio flag, como se espera.</summary>
         public static void EnsureAttDispNormal()
         {
-            AcApp.SetSystemVariable("ATTDISP", 1);
+            // ATTDISP es de tipo entero de 16 bits (short) a nivel interno: pasar un
+            // int (32 bits) sin mas -aunque el valor "1" sea valido- dispara
+            // eInvalidInput en SetSystemVariable. Hay que forzar el tipo exacto.
+            AcApp.SetSystemVariable("ATTDISP", (short)1);
         }
     }
